refactor: 유저-라운지 라운지 도메인에 포함하도록 수정
